+++ rq/decorators.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from collections.abc import Callable
+from collections.abc import Callable, Sequence
 from functools import wraps
 from typing import TYPE_CHECKING, Any
 
@@ -12,6 +12,7 @@ if TYPE_CHECKING:
 from .defaults import DEFAULT_RESULT_TTL
 from .job import Callback
 from .queue import Queue
+from .webhook import Webhook
 
 
 class job:  # noqa
@@ -34,6 +35,7 @@ class job:  # noqa
         on_failure: Callback | Callable[..., Any] | None = None,
         on_success: Callback | Callable[..., Any] | None = None,
         on_stopped: Callback | Callable[..., Any] | None = None,
+        webhooks: Sequence[Webhook] | None = None,
     ):
         """A decorator that adds a ``enqueue`` method to the decorated function,
         which in turn creates a RQ job when called. Accepts a required
@@ -68,6 +70,8 @@ class job:  # noqa
                 to None.
             on_stopped (Optional[Union[Callback, Callable[..., Any]]], optional): Callable to run when stopped. Defaults
                 to None.
+            webhooks (Optional[Sequence[Webhook]], optional): Webhooks to send on matching terminal job statuses.
+                Defaults to None.
         """
         self.queue = queue
         self.queue_class = queue_class if queue_class else Queue
@@ -84,6 +88,7 @@ class job:  # noqa
         self.on_success = on_success
         self.on_failure = on_failure
         self.on_stopped = on_stopped
+        self.webhooks = webhooks
 
     def __call__(self, f):
         @wraps(f)
@@ -120,6 +125,7 @@ class job:  # noqa
                 on_failure=self.on_failure,
                 on_success=self.on_success,
                 on_stopped=self.on_stopped,
+                webhooks=self.webhooks,
             )
 
         f.delay = delay  # TODO: Remove this in 3.0